backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.12",
    "markdown>=3.7",
    "passlib>=1.7.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
    "watchdog>=4.0.0",
]
